重新对城市进行编码
