+++ src/main/java/com/zzrb/enumm/CityIdEnum.java
@@ -1,396 +1,398 @@
 package com.zzrb.enumm;
 
 /**
- * description: CityIdEnum <br>
+ * description: CityCodeEnum <br>
  * date: 2020/11/5 3:52 下午 <br>
  * author: kuan <br>
- * version: 该枚举类弃用，请使用CityCodeEnum <br>
+ * version: 1.0 <br>
  */
-@Deprecated
 public enum CityIdEnum {
 
-    BEIJING("北京","010"),
-    TIANJIN("天津","022"),
-    SHIJIAZHUANG("石家庄","0311"),
-    TANGSHAN("唐山","0315"),
-    QINHUANGDAO("秦皇岛","0335"),
-    HANDAN("邯郸","0310"),
-    XINGTAI("邢台","0319"),
-    BAODING("保定","0312"),
-    ZHANGJIAKOU("张家口","0313"),
-    CHENGDE("承德","0314"),
-    CANGZHOU("沧州","0317"),
-    LANGFANG("廊坊","0316"),
-    HENGSHUI("衡水","0318"),
-    TAIYUAN("太原","0351"),
-    DATONG("大同","0352"),
-    YANGQUAN("阳泉","0353"),
-    CHANGZHI("长治","0355"),
-    JINCHENG("晋城","0356"),
-    SHUOZHOU("朔州","0349"),
-    JINZHONG("晋中","0354"),
-    YUNCHENG("运城","0359"),
-    XINZHOU("忻州","0350"),
-    LINFEN("临汾","0357"),
-    LVLIANG("吕梁","0358"),
-    HOHHOT("呼和浩特","0471"),
-    BAOTOU("包头","0472"),
-    WUHAI("乌海","0473"),
-    CHIFENG("赤峰","0476"),
-    TONGLIAO("通辽","0475"),
-    ORDOS("鄂尔多斯","0477"),
-    HULUNBER("呼伦贝尔","0470"),
-    BAYANNUR("巴彦淖尔","0478"),
-    ULANQAB("乌兰察布","0474"),
-    HINGGAN("兴安盟","0482"),
-    XILINGOL("锡林郭勒盟","0479"),
-    ALXA("阿拉善盟","0483"),
-    SHENYANG("沈阳","024"),
-    DALIAN("大连","0411"),
-    ANSHAN("鞍山","0412"),
-    FUSHUN("抚顺","024"),
-    BENXI("本溪","0414"),
-    DANDONG("丹东","0415"),
-    JINZHOU("锦州","0416"),
-    YINGKOU("营口","0417"),
-    FUXIN("阜新","0418"),
-    LIAOYANG("辽阳","0419"),
-    PANJIN("盘锦","0427"),
-    TIELING("铁岭","024"),
-    CHAOYANG("朝阳","0421"),
-    HULUDAO("葫芦岛","0429"),
-    JINPUXINQU("金普新区","0411"),
-    CHANGCHUN("长春","0431"),
-    JILIN("吉林","0432"),
-    SIPING("四平","0434"),
-    LIAOYUAN("辽源","0437"),
-    TONGHUA("通化","0435"),
-    BAISHAN("白山","0439"),
-    SONGYUAN("松原","0438"),
-    BAICHENG("白城","0436"),
-    YANBIAN("延边","0433"),
-    HARBIN("哈尔滨","0451"),
-    QIQIHAR("齐齐哈尔","0452"),
-    JIXI("鸡西","0467"),
-    HEGANG("鹤岗","0468"),
-    SHUANGYASHAN("双鸭山","0469"),
-    DAQING("大庆","0459"),
-    YICHUN("伊春","0458"),
-    JIANGXI_YICHUN("宜春","0795"),
-    JIAMUSI("佳木斯","0454"),
-    QITAIHE("七台河","0464"),
-    MUDANJIANG("牡丹江","0453"),
-    HEIHE("黑河","0456"),
-    SUIHUA("绥化","0455"),
-    DAXINGANLING("大兴安岭","0457"),
-    SHANGHAI("上海","021"),
-    NANJING("南京","025"),
-    WUXI("无锡","0510"),
-    XUZHOU("徐州","0516"),
-    CHANGZHOU("常州","0519"),
-    SUZHOU("苏州","0512"),
-    ANHUI_SUZHOU("宿州","0557"),
-    NANTONG("南通","0513"),
-    LIANYUNGANG("连云港","0518"),
-    HUAIAN("淮安","0517"),
-    YANCHENG("盐城","0515"),
-    YANGZHOU("扬州","0514"),
-    ZHENJIANG("镇江","0511"),
-    TAIZHOU("泰州","0523"),
-    ZHEJIANG_TAIZHOU("台州","0576"),
-    SUQIAN("宿迁","0527"),
-    HANGZHOU("杭州","0571"),
-    NINGBO("宁波","0574"),
-    WENZHOU("温州","0577"),
-    JIAXING("嘉兴","0573"),
-    HUZHOU("湖州","0572"),
-    SHAOXING("绍兴","0575"),
-    JINHUA("金华","0579"),
-    QUZHOU("衢州","0570"),
-    ZHOUSHAN("舟山","0580"),
-    LISHUI("丽水","0578"),
-    HEFEI("合肥","0551"),
-    WUHU("芜湖","0551"),
-    BENGBU("蚌埠","0552"),
-    HUAINAN("淮南","0554"),
-    MAANSHAN("马鞍山","0555"),
-    HUAIBEI("淮北","0561"),
-    TONGLING("铜陵","0562"),
-    ANQING("安庆","0556"),
-    HUANGSHAN("黄山","0559"),
-    CHUZHOU("滁州","0550"),
-    FUYANG("阜阳","0558"),
-    LUAN("六安","0564"),
-    BOZHOU("亳州","0558"),
-    CHIZHOU("池州","0566"),
-    XUANCHENG("宣城","0563"),
-    FUZHOU("福州","0591"),
-    JIANGXI_FUZHOU("抚州","0794"),
-    XIAMEN("厦门","0592"),
-    PUTIAN("莆田","0594"),
-    SANMING("三明","0598"),
-    QUANZHOU("泉州","0595"),
-    ZHANGZHOU("漳州","0596"),
-    NANPING("南平","0599"),
-    LONGYAN("龙岩","0597"),
-    NINGDE("宁德","0593"),
-    NANCHANG("南昌","0791"),
-    JINGDEZHEN("景德镇","0798"),
-    PINGXIANG("萍乡","0799"),
-    JIUJIANG("九江","0792"),
-    XINYU("新余","0790"),
-    YINGTAN("鹰潭","0701"),
-    GANZHOU("赣州","0797"),
-    JIAN("吉安","0796"),
-    SHANGRAO("上饶","0793"),
-    JINAN("济南","0531"),
-    QINGDAO("青岛","0532"),
-    ZIBO("淄博","0533"),
-    ZAOZHUANG("枣庄","0632"),
-    DONGYING("东营","0546"),
-    YANTAI("烟台","0635"),
-    WEIFANG("潍坊","0536"),
-    JINING("济宁","0537"),
-    TAIAN("泰安","0538"),
-    WEIHAI("威海","0631"),
-    RIZHAO("日照","0633"),
-    LAIWU("莱芜","0634"),
-    LINYI("临沂","0539"),
-    DEZHOU("德州","0534"),
-    LIAOCHENG("聊城","0635"),
-    BINZHOU("滨州","0543"),
-    HEZE("菏泽","0530"),
-    ZHENGZHOU("郑州","0371"),
-    KAIFENG("开封","0378"),
-    LUOYANG("洛阳","0379"),
-    PINGDINGSHAN("平顶山","0375"),
-    ANYANG("安阳","0372"),
-    HEBI("鹤壁","0392"),
-    XINXIANG("新乡","0373"),
-    JIAOZUO("焦作","0391"),
-    PUYANG("濮阳","0393"),
-    XUCHANG("许昌","0374"),
-    LUOHE("漯河","0395"),
-    SANMENXIA("三门峡","0398"),
-    NANYANG("南阳","0377"),
-    SHANGQIU("商丘","0370"),
-    XINYANG("信阳","0376"),
-    ZHOUKOU("周口","0394"),
-    ZHUMADIAN("驻马店","0396"),
-    WUHAN("武汉","027"),
-    HUANGSHI("黄石","0714"),
-    SHIYAN("十堰","0719"),
-    YICHANG("宜昌","0717"),
-    XIANGYANG("襄阳","0710"),
-    EZHOU("鄂州","0711"),
-    JINGMEN("荆门","0724"),
-    XIAOGAN("孝感","0712"),
-    JINGZHOU("荆州","0716"),
-    HUANGGANG("黄冈","0713"),
-    XIANNING("咸宁","0715"),
-    SUIZHOU("随州","0722"),
-    ENSHI("恩施","0718"),
-    CHANGSHA("长沙","0731"),
-    ZHUZHOU("株洲","0731"),
-    XIANGTAN("湘潭","0731"),
-    HENGYANG("衡阳","0734"),
-    SHAOYANG("邵阳","0739"),
-    YUEYANG("岳阳","0730"),
-    CHANGDE("常德","0736"),
-    ZHANGJIAJIE("张家界","0744"),
-    YIYANG("益阳","0737"),
-    CHENZHOU("郴州","0735"),
-    YONGZHOU("永州","0746"),
-    HUAIHUA("怀化","0745"),
-    LOUDI("娄底","0738"),
-    XIANGXI("湘西","0743"),
-    GUANGZHOU("广州","020"),
-    SHAOGUAN("韶关","0751"),
-    SHENZHEN("深圳","0755"),
-    ZHUHAI("珠海","0756"),
-    SHANTOU("汕头","0754"),
-    FOSHAN("佛山","0757"),
-    JIANGMEN("江门","0750"),
-    ZHANJIANG("湛江","0759"),
-    MAOMING("茂名","0668"),
-    ZHAOQING("肇庆","0758"),
-    HUIZHOU("惠州","0752"),
-    MEIZHOU("梅州","0753"),
-    SHANWEI("汕尾","0660"),
-    HEYUAN("河源","0762"),
-    YANGJIANG("阳江","0662"),
-    QINGYUAN("清远","0763"),
-    DONGGUAN("东莞","0769"),
-    ZHONGSHAN("中山","0760"),
-    CHAOZHOU("潮州","0768"),
-    JIEYANG("揭阳","0633"),
-    YUNFU("云浮","0766"),
-    NANNING("南宁","0771"),
-    LIUZHOU("柳州","0772"),
-    GUILIN("桂林","0773"),
-    WUZHOU("梧州","0774"),
-    BEIHAI("北海","0779"),
-    FANGCHENGGANG("防城港","0770"),
-    QINZHOU("钦州","0777"),
-    GUIGANG("贵港","0775"),
-    YULIN("玉林","0775"),
-    SHANXI_YULIN("榆林","0912"),
-    BAISE("百色","0776"),
-    HEZHOU("贺州","0774"),
-    HECHI("河池","0778"),
-    LAIBIN("来宾","0772"),
-    CHONGZUO("崇左","0771"),
-    HAIKOU("海口","0898"),
-    SANYA("三亚","08982"),
-    SANSHA("三沙","08983"),
-    CHONGQING("重庆","023"),
-    LIANGJIANGXINQU("两江新区","023"),
-    CHENGDU("成都","028"),
-    ZIGONG("自贡","0813"),
-    PANZHIHUA("攀枝花","0812"),
-    LUZHOU("泸州","0830"),
-    DEYANG("德阳","0838"),
-    MIANYANG("绵阳","0816"),
-    GUANGYUAN("广元","0839"),
-    SUINING("遂宁","0825"),
-    NEIJIANG("内江","0832"),
-    LESHAN("乐山","0833"),
-    NANCHONG("南充","0817"),
-    MEISHAN("眉山","028"),
-    YIBIN("宜宾","0831"),
-    GUANGAN("广安","0826"),
-    DAZHOU("达州","0818"),
-    YAAN("雅安","0835"),
-    BAZHONG("巴中","0827"),
-    ZIYANG("资阳","028"),
-    ABA("阿坝","0837"),
-    GARZE("甘孜","0836"),
-    LIANGSHAN("凉山","0834"),
-    GUIYANG("贵阳","0851"),
-    LIUPANSHUI("六盘水","0858"),
-    ZUNYI("遵义","0852"),
-    ANSHUN("安顺","0853"),
-    BIJIE("毕节","0857"),
-    TONGREN("铜仁","0856"),
-    QIANXINAN("黔西南","0859"),
-    QIANDONGNAN("黔东南","0855"),
-    QIANNAN("黔南","0854"),
-    KUNMING("昆明","0871"),
-    QUJING("曲靖","0874"),
-    YUXI("玉溪","0877"),
-    BAOSHAN("保山","0875"),
-    ZHAOTONG("昭通","0870"),
-    LIJIANG("丽江","0888"),
-    PUER("普洱","0879"),
-    LINCANG("临沧","0883"),
-    CHUXIONG("楚雄","0878"),
-    HONGHE("红河","0873"),
-    WENSHAN("文山","0876"),
-    XISHUANGBANNA("西双版纳","0691"),
-    DALI("大理","0872"),
-    DEHONG("德宏","0692"),
-    NUJIANG("怒江","0886"),
-    DEQEN("迪庆","0887"),
-    LHASA("拉萨","0891"),
-    RIKAZE("日喀则","0892"),
-    QAMDO("昌都","0895"),
-    SHANNAN("山南","0893"),
-    NAGQU("那曲","0896"),
-    NGARI("阿里","0897"),
-    NYINGCHI("林芝","0894"),
-    XIAN("西安","029"),
-    TONGCHUAN("铜川","0919"),
-    BAOJI("宝鸡","0917"),
-    XIANYANG("咸阳","029"),
-    WEINAN("渭南","0913"),
-    YANAN("延安","0911"),
-    HANZHONG("汉中","0916"),
-    ANKANG("安康","0915"),
-    SHANGLUO("商洛","0914"),
-    XIXIAN("西咸","029"),
-    LANZHOU("兰州","0931"),
-    JIAYUGUAN("嘉峪关","0937"),
-    JINCHANG("金昌","0935"),
-    BAIYIN("白银","0943"),
-    TIANSHUI("天水","0938"),
-    WUWEI("武威","0935"),
-    ZHANGYE("张掖","0936"),
-    PINGLIANG("平凉","0933"),
-    JIUQUAN("酒泉","0937"),
-    QINGYANG("庆阳","0934"),
-    DINGXI("定西","0932"),
-    LONGNAN("陇南","0939"),
-    LINXIA("临夏","0930"),
-    GANNAN("甘南","0941"),
-    XINING("西宁","0971"),
-    HAIDONG("海东","0972"),
-    HAIBEI("海北","0970"),
-    HUANGNAN("黄南","0973"),
-    HAINAN("海南","0974"),
-    GOLOG("果洛","0975"),
-    YUSHU("玉树","0976"),
-    HAIXI("海西","0977"),
-    YINCHUAN("银川","0951"),
-    SHIZUISHAN("石嘴山","0952"),
-    WUZHONG("吴忠","0953"),
-    GUYUAN("固原","0954"),
-    ZHONGWEI("中卫","0955"),
-    URUMQI("乌鲁木齐","0991"),
-    KARAMAY("克拉玛依","0990"),
-    TURPAN("吐鲁番","0995"),
-    HAMI("哈密","0902"),
-    CHANGJI("昌吉","0994"),
-    BORTALA("博尔塔拉","0909"),
-    BAYINGOL("巴音郭楞","0996"),
-    AKSU("阿克苏","0997"),
-    KIZILSU("克孜勒苏","0908"),
-    KASHGAR("喀什","0998"),
-    HOTAN("和田","0903"),
-    ILI("伊犁","0999"),
-    QOQEK("塔城","0901"),
-    ALTAY("阿勒泰","0906"),
-    TAIPEI("台北","02"),
-    KAOHSIUNG("高雄","07"),
-    KEELUNG("基隆","02"),
-    TAICHUNG("台中","04"),
-    TAINAN("台南","06"),
-    HSINCHU("新竹","03"),
-    CHIAYI("嘉义","05"),
-    NEWTAIPEI("新北","02"),
-    YILAN("宜兰","03"),
-    TAOYUAN("桃园","03"),
-    MIAOLI("苗栗","037"),
-    CHANGHUA("彰化","04"),
-    NANTOU("南投","049"),
-    YUNLIN("云林","05"),
-    PINGTUNG("屏东","08"),
-    TAITUNG("台东","089"),
-    HUALIEN("花莲","03"),
-    PENGHU("澎湖","06"),
-    JINMEN("金门","082"),
-    LIENCHIANG("连江","0836"),
-    HONGKONGISLAND("香港岛","00852"),
-    KOWLOON("九龙","00852"),
-    NEWTERRITORIES("新界","00852"),
-    MACAUPENINSULA("澳门半岛","00853"),
-    TAIPA("氹仔岛","00853"),
-    COLOANE("路环岛","00853"),
-    WUZHISHAN("五指山","089801"),
-    QIONGHAI("琼海","089802"),
-    DANZHOU("儋州","089803"),
-    WENCHANG("文昌","089805"),
-    WANNING("万宁","089806"),
-    DONGFANG("东方","089807"),
-    DINGAN("定安","089821"),
-    TUNCHANG("屯昌","089822"),
-    CHENGMAI("澄迈","089823"),
-    LINGAO("临高","089824"),
-    BAISHA("白沙","089825"),
-    CHANGJIANG("昌江","089826"),
-    LEDONG("乐东","089827"),
-    LINGSHUI("陵水","089828"),
-    BAOTING("保亭","089829"),
-    QIONGZHONG("琼中","089830");
+    BEIJING("北京","110100"),
+    TIANJIN("天津","120100"),
+    SHIJIAZHUANG("石家庄","130100"),
+    TANGSHAN("唐山","130200"),
+    QINHUANGDAO("秦皇岛","130300"),
+    HANDAN("邯郸","130400"),
+    XINGTAI("邢台","130500"),
+    BAODING("保定","130600"),
+    ZHANGJIAKOU("张家口","130700"),
+    CHENGDE("承德","130800"),
+    CANGZHOU("沧州","130900"),
+    LANGFANG("廊坊","131000"),
+    HENGSHUI("衡水","131100"),
+    TAIYUAN("太原","140100"),
+    DATONG("大同","140200"),
+    YANGQUAN("阳泉","140300"),
+    CHANGZHI("长治","140400"),
+    JINCHENG("晋城","140500"),
+    SHUOZHOU("朔州","140600"),
+    JINZHONG("晋中","140700"),
+    YUNCHENG("运城","140800"),
+    XINZHOU("忻州","140900"),
+    LINFEN("临汾","141000"),
+    LVLIANG("吕梁","141100"),
+    HOHHOT("呼和浩特","150100"),
+    BAOTOU("包头","150200"),
+    WUHAI("乌海","150300"),
+    CHIFENG("赤峰","150400"),
+    TONGLIAO("通辽","150500"),
+    ORDOS("鄂尔多斯","150600"),
+    HULUNBER("呼伦贝尔","150700"),
+    BAYANNUR("巴彦淖尔","150800"),
+    ULANQAB("乌兰察布","150900"),
+    HINGGAN("兴安盟","152200"),
+    XILINGOL("锡林郭勒盟","152500"),
+    ALXA("阿拉善盟","152900"),
+    SHENYANG("沈阳","210100"),
+    DALIAN("大连","210200"),
+    ANSHAN("鞍山","210300"),
+    FUSHUN("抚顺","210400"),
+    BENXI("本溪","210500"),
+    DANDONG("丹东","210600"),
+    JINZHOU("锦州","210700"),
+    YINGKOU("营口","210800"),
+    FUXIN("阜新","210900"),
+    LIAOYANG("辽阳","211000"),
+    PANJIN("盘锦","211100"),
+    TIELING("铁岭","211200"),
+    CHAOYANG("朝阳","211300"),
+    HULUDAO("葫芦岛","211400"),
+    JINPUXINQU("金普新区","211500"),
+    CHANGCHUN("长春","220100"),
+    JILIN("吉林","220200"),
+    SIPING("四平","220300"),
+    LIAOYUAN("辽源","220400"),
+    TONGHUA("通化","220500"),
+    BAISHAN("白山","220600"),
+    SONGYUAN("松原","220700"),
+    BAICHENG("白城","220800"),
+    YANBIAN("延边","222400"),
+    HARBIN("哈尔滨","230100"),
+    QIQIHAR("齐齐哈尔","230200"),
+    JIXI("鸡西","230300"),
+    HEGANG("鹤岗","230400"),
+    SHUANGYASHAN("双鸭山","230500"),
+    DAQING("大庆","230600"),
+    YICHUN("伊春","230700"),
+    JIAMUSI("佳木斯","230800"),
+    QITAIHE("七台河","230900"),
+    MUDANJIANG("牡丹江","231000"),
+    HEIHE("黑河","231100"),
+    SUIHUA("绥化","231200"),
+    DAXINGANLING("大兴安岭","232700"),
+    SHANGHAI("上海","310100"),
+    NANJING("南京","320100"),
+    WUXI("无锡","320200"),
+    XUZHOU("徐州","320300"),
+    CHANGZHOU("常州","320400"),
+    SUZHOU("苏州","320500"),
+    NANTONG("南通","320600"),
+    LIANYUNGANG("连云港","320700"),
+    HUAIAN("淮安","320800"),
+    YANCHENG("盐城","320900"),
+    YANGZHOU("扬州","321000"),
+    ZHENJIANG("镇江","321100"),
+    TAIZHOU("泰州","321200"),
+    SUQIAN("宿迁","321300"),
+    HANGZHOU("杭州","330100"),
+    NINGBO("宁波","330200"),
+    WENZHOU("温州","330300"),
+    JIAXING("嘉兴","330400"),
+    HUZHOU("湖州","330500"),
+    SHAOXING("绍兴","330600"),
+    JINHUA("金华","330700"),
+    QUZHOU("衢州","330800"),
+    ZHOUSHAN("舟山","330900"),
+    TAI_ZHOU("台州","331000"),
+    LISHUI("丽水","331100"),
+    ZHOU_SHAN("舟山新区","331200"),
+    HE_FEI("合肥","340100"),
+    WUHU("芜湖","340200"),
+    BENGBU("蚌埠","340300"),
+    HUAINAN("淮南","340400"),
+    MAANSHAN("马鞍山","340500"),
+    HUAIBEI("淮北","340600"),
+    TONGLING("铜陵","340700"),
+    ANQING("安庆","340800"),
+    HUANGSHAN("黄山","341000"),
+    CHUZHOU("滁州","341100"),
+    FUYANG("阜阳","341200"),
+    SU_ZHOU("宿州","341300"),
+    LUAN("六安","341500"),
+    BOZHOU("亳州","341600"),
+    CHIZHOU("池州","341700"),
+    XUANCHENG("宣城","341800"),
+    FUZHOU("福州","350100"),
+    XIAMEN("厦门","350200"),
+    PUTIAN("莆田","350300"),
+    SANMING("三明","350400"),
+    QUANZHOU("泉州","350500"),
+    ZHANGZHOU("漳州","350600"),
+    NANPING("南平","350700"),
+    LONGYAN("龙岩","350800"),
+    NINGDE("宁德","350900"),
+    NANCHANG("南昌","360100"),
+    JINGDEZHEN("景德镇","360200"),
+    PINGXIANG("萍乡","360300"),
+    JIUJIANG("九江","360400"),
+    XINYU("新余","360500"),
+    YINGTAN("鹰潭","360600"),
+    GANZHOU("赣州","360700"),
+    JIAN("吉安","360800"),
+    YI_CHUN("宜春","360900"),
+    FU_ZHOU("抚州","361000"),
+    SHANGRAO("上饶","361100"),
+    JINAN("济南","370100"),
+    QINGDAO("青岛","370200"),
+    ZIBO("淄博","370300"),
+    ZAOZHUANG("枣庄","370400"),
+    DONGYING("东营","370500"),
+    YANTAI("烟台","370600"),
+    WEIFANG("潍坊","370700"),
+    JINING("济宁","370800"),
+    TAIAN("泰安","370900"),
+    WEIHAI("威海","371000"),
+    RIZHAO("日照","371100"),
+    LAIWU("莱芜","371200"),
+    LINYI("临沂","371300"),
+    DEZHOU("德州","371400"),
+    LIAOCHENG("聊城","371500"),
+    BINZHOU("滨州","371600"),
+    HEZE("菏泽","371700"),
+    ZHENGZHOU("郑州","410100"),
+    KAIFENG("开封","410200"),
+    LUOYANG("洛阳","410300"),
+    PINGDINGSHAN("平顶山","410400"),
+    ANYANG("安阳","410500"),
+    HEBI("鹤壁","410600"),
+    XINXIANG("新乡","410700"),
+    JIAOZUO("焦作","410800"),
+    PUYANG("濮阳","410900"),
+    XUCHANG("许昌","411000"),
+    LUOHE("漯河","411100"),
+    SANMENXIA("三门峡","411200"),
+    NANYANG("南阳","411300"),
+    SHANGQIU("商丘","411400"),
+    XINYANG("信阳","411500"),
+    ZHOUKOU("周口","411600"),
+    ZHUMADIAN("驻马店","411700"),
+    WUHAN("武汉","420100"),
+    HUANGSHI("黄石","420200"),
+    SHIYAN("十堰","420300"),
+    YICHANG("宜昌","420500"),
+    XIANGYANG("襄阳","420600"),
+    EZHOU("鄂州","420700"),
+    JINGMEN("荆门","420800"),
+    XIAOGAN("孝感","420900"),
+    JINGZHOU("荆州","421000"),
+    HUANGGANG("黄冈","421100"),
+    XIANNING("咸宁","421200"),
+    SUIZHOU("随州","421300"),
+    ENSHI("恩施","422800"),
+    CHANGSHA("长沙","430100"),
+    ZHUZHOU("株洲","430200"),
+    XIANGTAN("湘潭","430300"),
+    HENGYANG("衡阳","430400"),
+    SHAOYANG("邵阳","430500"),
+    YUEYANG("岳阳","430600"),
+    CHANGDE("常德","430700"),
+    ZHANGJIAJIE("张家界","430800"),
+    YIYANG("益阳","430900"),
+    CHENZHOU("郴州","431000"),
+    YONGZHOU("永州","431100"),
+    HUAIHUA("怀化","431200"),
+    LOUDI("娄底","431300"),
+    XIANGXI("湘西","433100"),
+    GUANGZHOU("广州","440100"),
+    SHAOGUAN("韶关","440200"),
+    SHENZHEN("深圳","440300"),
+    ZHUHAI("珠海","440400"),
+    SHANTOU("汕头","440500"),
+    FOSHAN("佛山","440600"),
+    JIANGMEN("江门","440700"),
+    ZHANJIANG("湛江","440800"),
+    MAOMING("茂名","440900"),
+    ZHAOQING("肇庆","441200"),
+    HUIZHOU("惠州","441300"),
+    MEIZHOU("梅州","441400"),
+    SHANWEI("汕尾","441500"),
+    HEYUAN("河源","441600"),
+    YANGJIANG("阳江","441700"),
+    QINGYUAN("清远","441800"),
+    DONGGUAN("东莞","441900"),
+    ZHONGSHAN("中山","442000"),
+    CHAOZHOU("潮州","445100"),
+    JIEYANG("揭阳","445200"),
+    YUNFU("云浮","445300"),
+    NANNING("南宁","450100"),
+    LIUZHOU("柳州","450200"),
+    GUILIN("桂林","450300"),
+    WUZHOU("梧州","450400"),
+    BEIHAI("北海","450500"),
+    FANGCHENGGANG("防城港","450600"),
+    QINZHOU("钦州","450700"),
+    GUIGANG("贵港","450800"),
+    YULIN("玉林","450900"),
+    BAISE("百色","451000"),
+    HEZHOU("贺州","451100"),
+    HECHI("河池","451200"),
+    LAIBIN("来宾","451300"),
+    CHONGZUO("崇左","451400"),
+    HAIKOU("海口","460100"),
+    SANYA("三亚","460200"),
+    SANSHA("三沙","460300"),
+    WUZHISHAN("五指山","469001"),
+    QIONGHAI("琼海","469002"),
+    DANZHOU("儋州","469003"),
+    WENCHANG("文昌","469005"),
+    WANNING("万宁","469006"),
+    DONGFANG("东方","469007"),
+    DINGAN("定安","469021"),
+    TUNCHANG("屯昌","469022"),
+    CHENGMAI("澄迈","469023"),
+    LINGAO("临高","469024"),
+    BAISHA("白沙","469025"),
+    CHANGJIANG("昌江","469026"),
+    LEDONG("乐东","469027"),
+    LINGSHUI("陵水","469028"),
+    BAOTING("保亭","469029"),
+    QIONGZHONG("琼中","469030"),
+    CHONGQING("重庆","500100"),
+    LIANGJIANGXINQU("两江新区","500300"),
+    CHENGDU("成都","510100"),
+    ZIGONG("自贡","510300"),
+    PANZHIHUA("攀枝花","510400"),
+    LUZHOU("泸州","510500"),
+    DEYANG("德阳","510600"),
+    MIANYANG("绵阳","510700"),
+    GUANGYUAN("广元","510800"),
+    SUINING("遂宁","510900"),
+    NEIJIANG("内江","511000"),
+    LESHAN("乐山","511100"),
+    NANCHONG("南充","511300"),
+    MEISHAN("眉山","511400"),
+    YIBIN("宜宾","511500"),
+    GUANGAN("广安","511600"),
+    DAZHOU("达州","511700"),
+    YAAN("雅安","511800"),
+    BAZHONG("巴中","511900"),
+    ZIYANG("资阳","512000"),
+    ABA("阿坝","513200"),
+    GARZE("甘孜","513300"),
+    LIANGSHAN("凉山","513400"),
+    GUIYANG("贵阳","520100"),
+    LIUPANSHUI("六盘水","520200"),
+    ZUNYI("遵义","520300"),
+    ANSHUN("安顺","520400"),
+    BIJIE("毕节","520500"),
+    TONGREN("铜仁","520600"),
+    QIANXINAN("黔西南","522300"),
+    QIANDONGNAN("黔东南","522600"),
+    QIANNAN("黔南","522700"),
+    KUNMING("昆明","530100"),
+    QUJING("曲靖","530300"),
+    YUXI("玉溪","530400"),
+    BAOSHAN("保山","530500"),
+    ZHAOTONG("昭通","530600"),
+    LIJIANG("丽江","530700"),
+    PUER("普洱","530800"),
+    LINCANG("临沧","530900"),
+    CHUXIONG("楚雄","532300"),
+    HONGHE("红河","532500"),
+    WENSHAN("文山","532600"),
+    XISHUANGBANNA("西双版纳","532800"),
+    DALI("大理","532900"),
+    DEHONG("德宏","533100"),
+    NUJIANG("怒江","533300"),
+    DEQEN("迪庆","533400"),
+    LHASA("拉萨","540100"),
+    RIKAZE("日喀则","540200"),
+    QAMDO("昌都","540300"),
+    SHANNAN("山南","542200"),
+    NAGQU("那曲","542400"),
+    NGARI("阿里","542500"),
+    NYINGCHI("林芝","542600"),
+    XIAN("西安","610100"),
+    TONGCHUAN("铜川","610200"),
+    BAOJI("宝鸡","610300"),
+    XIANYANG("咸阳","610400"),
+    WEINAN("渭南","610500"),
+    YANAN("延安","610600"),
+    HANZHONG("汉中","610700"),
+    YU_LIN("榆林","610800"),
+    ANKANG("安康","610900"),
+    SHANGLUO("商洛","611000"),
+    XIXIAN("西咸","611100"),
+    LANZHOU("兰州","620100"),
+    JIAYUGUAN("嘉峪关","620200"),
+    JINCHANG("金昌","620300"),
+    BAIYIN("白银","620400"),
+    TIANSHUI("天水","620500"),
+    WUWEI("武威","620600"),
+    ZHANGYE("张掖","620700"),
+    PINGLIANG("平凉","620800"),
+    JIUQUAN("酒泉","620900"),
+    QINGYANG("庆阳","621000"),
+    DINGXI("定西","621100"),
+    LONGNAN("陇南","621200"),
+    LINXIA("临夏","622900"),
+    GANNAN("甘南","623000"),
+    XINING("西宁","630100"),
+    HAIDONG("海东","630200"),
+    HAIBEI("海北","632200"),
+    HUANGNAN("黄南","632300"),
+    HAINAN("海南","632500"),
+    GOLOG("果洛","632600"),
+    YUSHU("玉树","632700"),
+    HAIXI("海西","632800"),
+    YINCHUAN("银川","640100"),
+    SHIZUISHAN("石嘴山","640200"),
+    WUZHONG("吴忠","640300"),
+    GUYUAN("固原","640400"),
+    ZHONGWEI("中卫","640500"),
+    URUMQI("乌鲁木齐","650100"),
+    KARAMAY("克拉玛依","650200"),
+    TURPAN("吐鲁番","652100"),
+    HAMI("哈密","652200"),
+    CHANGJI("昌吉","652300"),
+    BORTALA("博尔塔拉","652700"),
+    BAYINGOL("巴音郭楞","652800"),
+    AKSU("阿克苏","652900"),
+    KIZILSU("克孜勒苏","653000"),
+    KASHGAR("喀什","653100"),
+    HOTAN("和田","653200"),
+    ILI("伊犁","654000"),
+    QOQEK("塔城","654200"),
+    ALTAY("阿勒泰","654300"),
+    TAIPEI("台北","710100"),
+    KAOHSIUNG("高雄","710200"),
+    KEELUNG("基隆","710300"),
+    TAICHUNG("台中","710400"),
+    TAINAN("台南","710500"),
+    HSINCHU("新竹","710600"),
+    CHIAYI("嘉义","710700"),
+    NEWTAIPEI("新北","710800"),
+    YILAN("宜兰","712200"),
+    TAOYUAN("桃园","712300"),
+    HS_INCHU("新竹","712400"),
+    MIAOLI("苗栗","712500"),
+    CHANGHUA("彰化","712700"),
+    NANTOU("南投","712800"),
+    YUNLIN("云林","712900"),
+    CHIA_YI("嘉义","713000"),
+    PINGTUNG("屏东","713300"),
+    TAITUNG("台东","713400"),
+    HUALIEN("花莲","713500"),
+    PENGHU("澎湖","713600"),
+    JINMEN("金门","713700"),
+    LIENCHIANG("连江","713800"),
+    HONGKONGISLAND("香港岛","810100"),
+    KOWLOON("九龙","810200"),
+    NEWTERRITORIES("新界","810300"),
+    MACAUPENINSULA("澳门半岛","820100"),
+    TAIPA("氹仔岛","820200"),
+    COLOANE("路环岛","820300");
 
     private String name;
     private String code;
@@ -413,7 +415,7 @@ public enum CityIdEnum {
     }
 
     // 构造方法
-    CityIdEnum(String name,String code) {
+    CityIdEnum(String name, String code) {
         this.name = name;
         this.code = code;
     }
